Change enclave to use handshake specific values based on SGX session mapping
This completes handling of concurrent SSL handshakes within the enclave.
Every SSL handshake that goes through the enclave is mapped against the
SGX session id set by the web server that handles the events. Each time
a command is executed within the enclave, a pointer is updated to show
at the current SGX session where the handshake specific variables exist.

Note: Yet, this mechanism does not render the enclave capable of
handling SSL session resumption because the SGX session id is set every
time a new connection is initiated by the web server and outside of the
SSL context (before receiving ClientRandom where the SSL session ID
might exist). Next commits will be on handling SSL sessions. SSL
SessionTickets is a different issue which will be tackled later on.